Moved thing to the auth module
