+++ app/src/main/java/com/apptec/registrateapp/auth/refreshtoken/RefreshTokenBody.java
@@ -1,4 +1,4 @@
-package com.apptec.registrateapp.repository.workers.retreshtoken;
+package com.apptec.registrateapp.auth.refreshtoken;
 
 public class RefreshTokenBody {
     /**
